chore: add CI workflow and clean up index test structure
- Add GitHub Actions CI that runs all tests with race detection on PRs
- Rename index/e2e_test.go to index/integration_test.go to reflect its
  actual purpose (search API integration tests, not HTTP e2e)
- Remove ~30 redundant tests already covered by unit tests in
  writer_test.go, disk_segment_test.go, merger_test.go
- Keep search API integration tests and bugfix regression guards
